+++ test/server_mongo.js
@@ -1,9 +1,10 @@
 var mqtt = require("mqtt");
-var async = require("async");
+var steed = require("steed");
 var ascoltatori = require("ascoltatori");
 var abstractServerTests = require("./abstract_server");
-var MongoClient = require('mongodb').MongoClient;
+var MongoClient = require("mongodb").MongoClient;
 var clean = require("mongo-clean");
+var createConnection = require("./helpers/createConnection");
 
 describe("mosca.Server with mongo persistence", function() {
   this.timeout(10000);
@@ -32,12 +33,12 @@ describe("mosca.Server with mongo persistence", function() {
       port: nextPort(),
       stats: false,
       publishNewClient: false,
+      publishClientDisconnect: false,
       logger: {
-        childOf: globalLogger,
-        level: 60
+        level: "error"
       },
       backend : {
-        type: 'mongo'
+        type: "mongo"
         // not reusing the connection
         // because ascoltatori has not an autoClose option
         // TODO it must be handled in mosca.Server
@@ -50,5 +51,5 @@ describe("mosca.Server with mongo persistence", function() {
     };
   }
 
-  abstractServerTests(moscaSettings, mqtt.createConnection);
+  abstractServerTests(moscaSettings, createConnection);
 });